+++ lib/tabs/account.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class AccountPage extends StatelessWidget {
+  const AccountPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Placeholder();
+  }
+}
